+++ LuckyoneUI/Core/Core.lua
@@ -493,4 +493,8 @@ function Private.Addon:RegisterEvents()
 	self:RegisterEvent('PLAYER_ENTERING_WORLD')
 	self:RegisterEvent('PLAYER_LOGIN')
 	self:RegisterEvent('PLAYER_SPECIALIZATION_CHANGED')
+
+	if Private.ElvUI then
+		Private:MiscTweaks()
+	end
 end